nixos/anbox: add --daemon flag
This supresses the following warning in anbox-container-manager:

You are running the container manager manually which is most likely not
what you want. The container manager is normally started by systemd or
another init system. If you still want to run the container-manager
you can get rid of this warning by starting with the --daemon option.