smartmontools: disable mail support by default
Mail support only really needs to be enabled when smartmontools is used as a
daemon and even then it's only really needed if the daemon is configured to
actually send mails.

For enabling mails, smartmontools depends on guile which takes forever to
build (especially important on platforms without binary cache) and bloats the
closure size.

Closure size: 199634368 -> 50533248 (200MB -> 50MB)

Since smartmontools are in nixos/base, the closure size of the system drv inside
the minimal ISO also decreases considerably:

2381629752 -> 2316287728 (2.38GB -> 2.31GB)