ffmpeg: make version part of the interface
This allows overriders to declare a different ABI. Doing so via overrideAttrs
would not affect the flags in the arguments; effectively retaining the
overridden package's ABI.
See https://github.com/NixOS/nixpkgs/issues/280645 for an instance of that.

By overriding the arguments using

    ffmpeg.override { version = "..."; ... }

the ABI will now be overridden as expected.